# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hospital Management API",
        "description": "API documentation for Hospital, Doctor, Patient, and Appointment management system",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/patients": {
            "get": {
                "tags": [
                    "patient"
                ],
                "summary": "Get all patients",
                "operationId": "6747f925788f494507252b8889e09dd1",
                "responses": {
                    "200": {
                        "description": "List of patient"
                    }
                }
            },
            "post": {
                "tags": [
                    "patient"
                ],
                "summary": "Create a new patient",
                "operationId": "9296610ded1a685f9284d36de20ab313",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "address",
                                    "DOB",
                                    "gender",
                                    "hospital_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "DOB": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "format": "enum"
                                    },
                                    "hospital_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Patient created successfully"
                    }
                }
            }
        },
        "/api/v1/patients/{id}": {
            "get": {
                "tags": [
                    "patient"
                ],
                "summary": "Get a Patient by ID",
                "operationId": "054f2ccead52a541c84d2434da85858d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Patient",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient data retrieved"
                    },
                    "404": {
                        "description": "Patient not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "patient"
                ],
                "summary": "Update a patient",
                "operationId": "3b8d0a85cd7984d11ab97dea48582496",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the patient to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "address",
                                    "DOB",
                                    "gender",
                                    "hospital_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "DOB": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "format": "enum"
                                    },
                                    "hospital_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Patient updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "patient"
                ],
                "summary": "Delete a Patient",
                "operationId": "593eda9a73fd98782af771e44d1adef6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Patient to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Patient deleted successfully"
                    }
                }
            }
        }
    },
...
}